Tests für Overtaking
